analytics methods with progress
more done with analytics methods

done with analytics

+++ db/migrate/20190507092235_create_price_records.rb
@@ -4,7 +4,7 @@ class CreatePriceRecords < ActiveRecord::Migration[5.2]
       t.datetime :effective_from
       t.float :monthly_price
       t.float :joining_fee
-      t.float :service_id
+      t.integer :service_id
       t.timestamps
     end
   end
